Modified rob_traffic_light to enable multiple language use.
Added three parameters to the rob_traffic_light function:
* x_title which is passed to scale_x_discrete in the generic template

* y_title which is passed to scale_y_continuous in the generic template

* judgement_labels which are passed to scale_colour_manual and
are named in line with judgement_levels, updated @params for these
specifying that judgement_labels need to be in the proscribed order.

Modified the assignment of judgement levels to create a new variable
(judgement_levels) with which to name judgement_labels.